backend/requirements2.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.1
greenlet==3.1.0
h11==0.14.0
idna==3.8
numpy==2.1.1
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
